module2/lesson2_1/apps/order/api_endpoints/ProductList/serializers.py
from rest_framework import serializers

from apps.order.models import Product

class ProductListSerializer(serializers.ModelSerializer):

    class Meta:
        model = Product
        fields = ("id", "name", "price", "is_active", "image")