module2/lesson2_1/requirements.txt
Pillow==11.3.0
django==4.2.23
djangorestframework==3.16.0